// iterpro-frontend/libs/shared/data-access/sdk/src/models/interfaces/etl/etlDuration.ts
export enum EtlDateType {
	MICROSOFT_TIME = 0,
	UNIX_TIME = 1,
	DDMMYYYY = 2,
	DDMMYY = 3,
	MICROSOFT_TIME_FLOAT = 4,
	HHMMSS_NODATE = 5,
	YYYYMMDDHHMMSS = 6,
	DDMMYYYYHHMMSS = 7,
	MMDDYYYY = 8
}

export enum EtlDurationType {
	SECONDS = 0,
	HHMMSS = 1,
	MMSS = 4,
	MINUTES = 2,
	MINUTESFLOAT = 3,
	EXCELDECIMALSECONDS = 5
}